build udev lib bins statically
This patch extracts both stratis-base32-decode and stratis-str-cmp from
stratis-util. They are now two distinct binaries.
Also, when built, they are built statically.

```
$ ldd target/release/stratis-str-cmp
        statically linked
```

Related: https://github.com/stratis-storage/stratisd/issues/3195
