+++ .github/workflows/fedora.yml
@@ -109,6 +109,7 @@ jobs:
           curl
           device-mapper-persistent-data
           dbus-devel
+          glibc-static
           libblkid-devel
           make
           systemd-devel
