+++ .github/workflows/python.yml
@@ -55,6 +55,7 @@ jobs:
           dbus-devel
           dbus-tools
           device-mapper-persistent-data
+          glibc-static
           libblkid-devel
           make
           python3-dbus
@@ -120,6 +121,7 @@ jobs:
           dbus-devel
           dbus-tools
           device-mapper-persistent-data
+          glibc-static
           libblkid-devel
           make
           python3-dbus
